+++ src/main/java/org/la4j/matrix/Matrix.java
@@ -982,13 +982,60 @@ public interface Matrix extends Externalizable {
     double foldColumn(int j, MatrixAccumulator accumulator);
 
     /**
-     * Folds all elements (in column-by-column manner) of this matrix with given {@code accumulator}.
+     * Folds all elements (in a column-by-column manner) of this matrix with given {@code accumulator}.
      *
      * @param accumulator the matrix accumulator
      *
      * @return the accumulated vector
      */
     Vector foldColumns(MatrixAccumulator accumulator);
+
+    /**
+     * Folds non-zero elements of this matrix with given {@code accumulator}.
+     *
+     * @param accumulator the matrix accumulator
+     *
+     * @return the accumulated value
+     */
+    double foldNonZero(MatrixAccumulator accumulator);
+
+    /**
+     * Folds non-zero elements of specified row in this matrix with given {@code accumulator}.
+     *
+     * @param i the row index
+     * @param accumulator the matrix accumulator
+     *
+     * @return the accumulated value
+     */
+    double foldNonZeroInRow(int i, MatrixAccumulator accumulator);
+
+    /**
+     * Folds non-zero elements of specified column in this matrix with given {@code accumulator}.
+     *
+     * @param j the column index
+     * @param accumulator the matrix accumulator
+     *
+     * @return the accumulated value
+     */
+    double foldNonZeroInColumn(int j, MatrixAccumulator accumulator);
+
+    /**
+     * Folds non-zero elements (in a column-by-column manner) of this matrix with given {@code accumulator}.
+     *
+     * @param accumulator the matrix accumulator
+     *
+     * @return the accumulated vector
+     */
+    Vector foldNonZeroInColumns(MatrixAccumulator accumulator);
+
+    /**
+     * Folds non-zero elements (in a row-by-row manner) of this matrix with given {@code accumulator}.
+     *
+     * @param accumulator the matrix accumulator
+     *
+     * @return the accumulated vector
+     */
+    Vector foldNonZeroInRows(MatrixAccumulator accumulator);
 
     /**
      * Checks whether this matrix compiles with given {@code predicate} or not.