+++ CHANGELOG.md
@@ -2,6 +2,8 @@ Changelog
 ---------
 la4j-0.5.0 `May 2014`
  * Bug fix in serialization algorithm of sparse matrices (contributed by Catherine da Graca)`
+ * New matrix methods: `foldNonZero`, `foldNonZeroInRow`, `foldNonZeroInColumn`, `foldNonZeroInColumns`, `foldNonZeroInRows`
+ * New vector method: `foldNonZero`
 
 la4j-0.4.9 `Jan 2014`
  * Bug fix in `align()` method for big sparse matrices (reported by Michael Kapper)